% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summaries.R
\name{calculate_paucs.qpc}
\alias{calculate_paucs.qpc}
\title{calculate partial aucs for those defined in pauc}
\usage{
\method{calculate_paucs}{qpc}(x, rdv_time, rdv, sdv_time = rdv_time,
  sdv = rdv, id, rep)
}
\description{
calculate partial aucs for those defined in pauc
}

