% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summaries.R
\name{calculate_quantiles.qpc}
\alias{calculate_quantiles.qpc}
\title{get quantile summaries for real and simulated data for qpc object}
\usage{
\method{calculate_quantiles}{qpc}(x, probs, rep = "REP", na.rm = T)
}
\arguments{
\item{x}{qpc object}

\item{probs}{probabilities for which quantiles to calculate (eg 0.75 for 75th percentile)}

\item{rep}{column name for replicates in simulated data}
}
\description{
get quantile summaries for real and simulated data for qpc object
}
\examples{
\dontrun{
calculate_quantiles(qpcdb, probs = c(0.25, 0.5, 0.75))
}
}

