% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qpc.R
\docType{data}
\name{obs_data1}
\alias{obs_data1}
\title{simulated 'real data' for use in qpc vignette
\itemize{
  \item ID Numerical ID  (1--50)
  \item TIME Time in hours (0--24)
  \item DV Plasma concentration in mg/L
  \item AMT Amount of drug given, time dependent, in milligrams
  \item DOSE Dose given to each individual in milligrams
  \item MDV Missing dependent variable
  \item ISM is male
}}
\usage{
data(obs_data1)
}
\description{
simulated 'real data' for use in qpc vignette
\itemize{
  \item ID Numerical ID  (1--50)
  \item TIME Time in hours (0--24)
  \item DV Plasma concentration in mg/L
  \item AMT Amount of drug given, time dependent, in milligrams
  \item DOSE Dose given to each individual in milligrams
  \item MDV Missing dependent variable
  \item ISM is male
}
}
\keyword{datasets}

