% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qpc.R
\docType{data}
\name{sim_data1}
\alias{sim_data1}
\title{simulated data from 'real data' fit for obs_data1
\itemize{
  \item ID Numerical ID  (1--50)
  \item REP replicate
  \item TIME Time in hours (0--24)
  \item DV Plasma concentration in mg/L
  \item IPRED individual predicted value
  \item PRED population predicted value
  \item AMT Amount of drug given, time dependent, in milligrams
  \item DOSE Dose given to each individual in milligrams
  \item ISM is male
  \item MDV Missing dependent variable
}}
\usage{
data(sim_data1)
}
\description{
simulated data from 'real data' fit for obs_data1
\itemize{
  \item ID Numerical ID  (1--50)
  \item REP replicate
  \item TIME Time in hours (0--24)
  \item DV Plasma concentration in mg/L
  \item IPRED individual predicted value
  \item PRED population predicted value
  \item AMT Amount of drug given, time dependent, in milligrams
  \item DOSE Dose given to each individual in milligrams
  \item ISM is male
  \item MDV Missing dependent variable
}
}
\keyword{datasets}

